refactor: Update commit-msg and pre-commit hooks
This commit updates the commit-msg and pre-commit hooks in the .husky directory. The changes modify the mode of the files and remove the unnecessary newline at the end of the commit-msg file. The purpose of this update is to ensure that the commit messages are validated using commitlint and the lint-staged script is executed before committing the changes.